Now using float64 for timing for greater precision

+++ core/hitcounter/concurrent_map.go
@@ -9,7 +9,7 @@ const NumShards = 256
 
 type BlockStatus struct {
 	IsBlocked bool
-	FrontTile int32
+	FrontTile float64
 }
 
 // StringMap
